move fixture to correct folder
